# app/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code, all api start with "url:5000/api/<name>", for example "http://localhost:5000/api/user"
  version: "1.0.0"
  title: Swagger ReST Article
consumes:
  - application/json
produces:
  - application/json
basePath: /api
...
paths:
  /user:
    get:
      operationId: api.user.read_list
      tags:
        - User
      summary: Read list of users
      description: Read list of N users if user that call this API is not logged, otherwise return the first N users giving importance to users near the user that call this API. For this reason we pass to this API also the id of user that call it
      parameters:
      - name: token
        in: header
        description: Token generated for user during login
        required: False
        type: string
      - name: id
        in: header
        description: Id that identify user
        required: False
        type: integer
      - name: language
        in: query
        description: Language using in client
        required: True
        type: string
      responses:
        200:
          description: Successfully read people list operation
          schema:
            type: object
            properties:
              status: 
                type: integer
                description: status returned by api
              users:
                type: array
                description: list of N users with basic information show in travelers page
                items:
                  type: object
                  properties:
                    id: 
                      type: integer
                      description: Id of user
                    username: 
                      type: string
                      description: username of user
                    image: 
                      type: string
                      description: path of profile image of user
                    languages:
                      type: array
                      items:
                        type: object
                        properties:
                          language:
                            type: string
                            description: name of the language
                    home:
                      type: object
                      properties:
                        country:
                          type: string
                          description: country of locality visited by user
                        provice:
                          type: string
                          description: provice of locality visited by user
                        city:
                          type: string
                          description: city of locality visited by user
                
    post:
      operationId: api.user.create
      tags:
        - User
      summary: Create a new user
      description: Create a new user with basic information
      parameters:
        - name: language
          in: query
          description: Language using in client
          required: True
          type: string
        - name: data
          in: body
          description: all basic datas for the creation of a new user
          required: True
          schema:
            type: object
            properties:
              email:
                type: string
              username:
                type: string
              password:
                type: string
      responses:
        201:
          description: Successfully created new user
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
  
    delete:
      operationId: api.user.delete
      tags:
        - User
      summary: Delete 1 user
      description: Delete the user with the id passed
      parameters:
        - name: language
          in: query
          description: Language using in client
          required: True
          type: string
        - name: token
          in: header
          type: string
          description: Token created when user logged, send to verify if it is valid
          required: True
      responses:
        200: 
          description: Successfully deleted person from user table
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
  /user/profileImage:
    post:
      operationId: api.user.update_profile_image
      tags:
        - User
      summary: Update profile image
      description: Update profile image, if nothing is passed it just remove the current profile image
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: data
          in: body
          description: Image in base64 format
          required: False
          schema:
            type: object
            properties:
              image:
                type: string
                description: Image in base64 format
      responses:
        200:
          description: Successfully updated profile image
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
              image:
                type: string
                description: name of image and extension
  /user/{id}:
    get:
      operationId: api.user.read_all_details
      tags:
        - User
      summary: Details about 1 user
      description: Read all details about 1 single user
      parameters:
        - name: language
          in: query
          description: Language using in client
          required: True
          type: string
        - name: id
          in: path
          description: id that identify an user inside the DB I want to get details
          type: integer
          required: True
      responses:
        200:
          description: Successfully read user details
          schema:
            type: object
            properties:
              status: 
                type: integer
                description: status returned by api
              user:
                type: object
                properties:
                  id: 
                    type: integer
                    description: Id of user
                  username: 
                    type: string
                    description: username of user
                  image: 
                    type: string
                    description: path of profile image of user
                  experiences:
                    type: integer
                    description: number of experiences created by the user
                  itineraries:
                    type: integer
                    description: number of itineraries created by the user
                  matchings:
                    type: integer
                    description: number of matchings user did
                  tips:
                    type: integer
                    description: number of tips created by the user
                  home:
                    type: object
                    properties:
                      country:
                        type: string
                        description: country of locality visited by user
                      provice:
                        type: string
                        description: provice of locality visited by user
                      city:
                        type: string
                        description: city of locality visited by user
                  languages:
                    type: array
                    description: list of languages user can speak
                    items:
                      type: object
                      properties:
                        language:
                          type: string
                          description: name of the language
                  localities:
                    type: array
                    description: array of localities visited by the user
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: id that identify a locality
                        country:
                          type: string
                          description: country of locality visited by user
                        provice:
                          type: string
                          description: provice of locality visited by user
                        city:
                          type: string
                          description: city of locality visited by user
                  interests:
                    type: array
                    description: array of the interests of the user
                    items:
                      type: object
                      properties:
                        interest:
                          type: string
                          description: name of the interest
...
  /user/interest:
    put:
      operationId: api.user.put_interest
      tags:
        - User
      summary: Add 1 interest 
      description: Add 1 interest to the user that call api
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: interest
          in: query
          description: Interest we want to add
          type: string
          required: True
      responses:
        200:
          description: Successfully added new interest
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
  

    delete:
      operationId: api.user.delete_interest
      tags:
        - User
      summary: Delete 1 interest 
      description: Delete 1 interest to the user that call api
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: interest
          in: query
          description: Interest we want to delete
          type: string
          required: True
      responses:
        200: 
          description: Successfully removed interest
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
...
  /user/recoverPassword:
    get:
      operationId: api.user.recover_password
      tags:
        - User
      summary: Recover password
      description: Send an email to reset the password
      parameters:
        - name: email
          in: query
          type: string
          description: Email we want recover the password
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
      responses:
        200:
          description: Successfully send the email to reset the password
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
...